fix shadcn version and add landing page component
